+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 ## [0.3.1]
 * Performance improvements.
-* BREAK: Now uses `Point<int>` instead of Offset
+* BREAK: Now uses `(int, int)` instead of Offset
 
 ## [0.3.0]
 * Adds `AStar.byFreeSpaces`.